+++ romancal/flux/__init__.py
@@ -1,0 +1,3 @@
+from .flux_step import FluxStep
+
+__all__ = ["FluxStep"]